Fix pg_pk on tables not owned by user.
Current implementation of pg_pk uses
information_schema.constraint_column_usage, which only lists tables that
are owned by (a role of) the current user. [1]

This commit changes this to use the pg_attribute and pg_index views,
which also works if the user is not the owner of the table.

[1]: http://www.postgresql.org/docs/current/static/infoschema-constraint-column-usage.html